fix: tsconfig.json 및 파일명 수정
