+++ src/components/views/Community.tsx
@@ -1,23 +1,21 @@
 /** @jsxImportSource @emotion/react */
-import { css } from "@emotion/react";
+import { css } from '@emotion/react';
 import Calendar from './community/Calendar';
-import {Props} from './community/DateProps';
+import { Props } from './community/DateProps';
 
 const Community = () => {
-
-  const props:Props = {
+  const props: Props = {
     year: new Date().getFullYear(),
     month: new Date().getMonth(),
-  }
+  };
 
   return (
-   <div css={community}>
-      <div css={board}>
-      </div>
+    <div css={community}>
+      <div css={board}></div>
       <div className="calendar" css={calendar}>
-        <Calendar year={props.year} month={props.month}/>
+        <Calendar year={props.year} month={props.month} />
       </div>
-   </div>
+    </div>
   );
 };
